Migration of videoInfo to separate module
betteReddit refactor #906
